Pago con stripe funcional

+++ app/src/main/java/com/example/persistencia/Modelos/ProductoCarrito.kt
@@ -3,5 +3,5 @@ package com.example.persistencia.Modelos
 // Representación de un producto en el carrito de la compra
 data class ProductoCarrito(
     val producto: Producto, // El producto en sí
-    val cantidad: Int // La cantidad que se ha añadido
+    val cantidad: Double // La cantidad que se ha añadido
 )
